:truck: Move `docs_src` utils to general test utils
