+++ tests/docs_src/utils.py
@@ -1,37 +1,12 @@
 import importlib
-import json
 from types import ModuleType
-from typing import Dict, List, Optional
+from typing import List
 
-import orjson
-
-# Invoking tests on the command line (e.g. in CI) will capture logs directly
-# so they must be redirected back to sys.out/sys.err
-# Invoking test via IDE (i.e., Pycharm) sends logs directly to sys.out/sys.err
-from _pytest.capture import CaptureFixture
-from _pytest.logging import LogCaptureFixture
 from _pytest.monkeypatch import MonkeyPatch
 
 import tests.utils
 
-JSONOutputType = Dict[str, Optional[str]]
-
-
-def redirect_captured_logs_to_stdout(caplog: LogCaptureFixture) -> None:
-    if caplog.records:
-        for record in caplog.records:
-            print(orjson.dumps(record.msg).decode())
-
-
-def get_validated_json_output(capsys: CaptureFixture) -> List[JSONOutputType]:
-    captured = capsys.readouterr()
-    assert captured.out
-    assert not captured.err
-    out_str_loads_compatible = captured.out.strip().split("\n")
-    json_lib_ver = [json.loads(jsoned) for jsoned in out_str_loads_compatible]
-    orjson_lib_ver = [orjson.loads(jsoned) for jsoned in out_str_loads_compatible]
-    assert orjson_lib_ver == json_lib_ver
-    return orjson_lib_ver
+JSONOutputType = tests.utils.JSONOutputType
 
 
 def validate_output(
