+++ tests/utils.py
@@ -1,11 +1,15 @@
-from typing import List
+import json
+from typing import Any, Dict, List
 
 import orjson
 import structlog
 from _pytest.capture import CaptureFixture
+from _pytest.logging import LogCaptureFixture
 from _pytest.monkeypatch import MonkeyPatch
 
 import structlog_sentry_logger
+
+JSONOutputType = Dict[str, Any]
 
 
 def reset_logging_configs() -> None:
@@ -23,6 +27,26 @@ def enable_sentry_integration_mode(monkeypatch: MonkeyPatch) -> None:
         "STRUCTLOG_SENTRY_LOGGER_CLOUD_SENTRY_INTEGRATION_MODE_ON",
         "ANY_VALUE",
     )
+
+
+# Invoking tests on the command line (e.g. in CI) will capture logs directly
+# so they must be redirected back to sys.out/sys.err
+# Invoking test via IDE (i.e., Pycharm) sends logs directly to sys.out/sys.err
+def redirect_captured_logs_to_stdout(caplog: LogCaptureFixture) -> None:
+    if caplog.records:
+        for record in caplog.records:
+            print(orjson.dumps(record.msg).decode())
+
+
+def get_validated_json_output(capsys: CaptureFixture) -> List[JSONOutputType]:
+    captured = capsys.readouterr()
+    assert captured.out
+    assert not captured.err
+    out_str_loads_compatible = captured.out.strip().split("\n")
+    json_lib_ver = [json.loads(jsoned) for jsoned in out_str_loads_compatible]
+    orjson_lib_ver = [orjson.loads(jsoned) for jsoned in out_str_loads_compatible]
+    assert orjson_lib_ver == json_lib_ver
+    return orjson_lib_ver
 
 
 def read_json_logs_from_stdout(capsys: CaptureFixture) -> List[dict]:
